Touchbar toggle persistence, Pro input mode fix, Num inputer UI fix, and experimental Midi over UART

+++ src/Components/UIelement.h
@@ -9,6 +9,7 @@
 // #include "../Protocol/MIDI.h"
 #include "../HAL/KeyPad.h"
 #include "../HAL/LED.h"
+#include "../Components/GFX.h"
 // #include "../HAL/Timer.h"
 
 class UIelement
@@ -19,7 +20,7 @@ public:
   void renderHalfHeightNum(u16 num, u8 xy, u32 color, u32 sec_color);
   void renderHalfHeightDigit(u8 num, u8 xy, u32 color);
   u8 binary8bitInput(u8 currentNum, u8 y, u32 color);
-  u8 simple8bitInput(s16 currentNum, u8 y, u32 color);
+  u8 simple8bitInput(u8 currentNum, u8 y, u32 color);
 private:
 };
